Phase H: Edge Function fixed + Firestore sync worker + DB migrations
- Rewrote supabase/functions/order-lifecycle/index.ts
  - Removed firebase-admin (was causing BOOT_ERROR at runtime)
  - /health endpoint now zero-dependency (instant boot)
  - Supabase client lazily imported via esm.sh only for auth'd routes
  - Status mutations write to outbox_events table (outbox pattern)
  - Verified live: health check passing, auth gating working

- Added backend/src/services/firestore-sync-worker.js
  - Polls outbox_events every 5s, syncs to Firestore
  - Handles: order_status_changed, order_delivered, order_created,
    order_cancelled, inventory_updated, payment_completed
  - MAX_RETRY=3 with per-event error tracking

- Updated backend/render.yaml
  - Added fufaji-firestore-sync as a Render background worker

- DB migrations applied
  - 09_checkout_transactions.sql
  - 20260704120000_phase_h_critical_fixes.sql (OTP hash, ENUM, outbox)
  - 20260704130000_outbox_retry_rpc.sql (increment_retry_count RPC)

- Quarantined legacy Firebase functions to functions/src/_deprecated/

Architecture: Flutter -> Supabase Edge Functions -> Postgres -> outbox_events -> Render Sync Worker -> Firestore

+++ lib/models/coupon.dart
@@ -59,7 +59,7 @@ class Coupon {
       'name': name,
       'description': description,
       'discountType': discountType,
-      'discountValue': discountValue,
+      'discountValue': discountValue.toDouble(),
       'minimumOrderAmount': minimumOrderAmount,
       'maximumDiscountAmount': maximumDiscountAmount,
       'startDate': Timestamp.fromDate(startDate),
